Move Estimator Classes to folder
