+++ src/rolch/estimators/__init__.py
@@ -1,0 +1,4 @@
+from .online_gamlss import OnlineGamlss
+from .online_lasso import OnlineLasso
+
+__all__ = ["OnlineGamlss", "OnlineLasso"]
